+++ src/main/java/com/telepathicgrunt/repurposedstructures/RSFeatures.java
@@ -45,17 +45,17 @@ public class RSFeatures {
     public static Feature<DefaultFeatureConfig> FORTRESS_BREAKAGE = new FortressBreakage(DefaultFeatureConfig.CODEC);
     public static Feature<DefaultFeatureConfig> STRONGHOLD_CHAINS = new StrongholdChains(DefaultFeatureConfig.CODEC);
 
-    public static StructureFeature<DefaultFeatureConfig> BIRCH_MINESHAFT = new RSMineshaftBirchStructure(DefaultFeatureConfig.CODEC);
-    public static StructureFeature<DefaultFeatureConfig> DESERT_MINESHAFT = new RSMineshaftDesertStructure(DefaultFeatureConfig.CODEC);
-    public static StructureFeature<DefaultFeatureConfig> END_MINESHAFT = new RSMineshaftEndStructure(DefaultFeatureConfig.CODEC);
-    public static StructureFeature<DefaultFeatureConfig> NETHER_MINESHAFT = new RSMineshaftNetherStructure(DefaultFeatureConfig.CODEC);
-    public static StructureFeature<DefaultFeatureConfig> ICY_MINESHAFT = new RSMineshaftIcyStructure(DefaultFeatureConfig.CODEC);
-    public static StructureFeature<DefaultFeatureConfig> JUNGLE_MINESHAFT = new RSMineshaftJungleStructure(DefaultFeatureConfig.CODEC);
-    public static StructureFeature<DefaultFeatureConfig> OCEAN_MINESHAFT = new RSMineshaftOceanStructure(DefaultFeatureConfig.CODEC);
-    public static StructureFeature<DefaultFeatureConfig> SAVANNA_MINESHAFT = new RSMineshaftSavannaStructure(DefaultFeatureConfig.CODEC);
-    public static StructureFeature<DefaultFeatureConfig> STONE_MINESHAFT = new RSMineshaftStoneStructure(DefaultFeatureConfig.CODEC);
-    public static StructureFeature<DefaultFeatureConfig> SWAMP_OR_DARK_FOREST_MINESHAFT = new RSMineshaftSwampOrDarkForestStructure(DefaultFeatureConfig.CODEC);
-    public static StructureFeature<DefaultFeatureConfig> TAIGA_MINESHAFT = new RSMineshaftTaigaStructure(DefaultFeatureConfig.CODEC);
+    public static StructureFeature<DefaultFeatureConfig> BIRCH_MINESHAFT = new RSMineshaftStructure(DefaultFeatureConfig.CODEC, RSMineshaftPieces.Type.BIRCH, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.spawnrate.birchMineshaftSpawnrate, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.maxHeight.birchMineshaftMaxHeight, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.minHeight.birchMineshaftMinHeight);
+    public static StructureFeature<DefaultFeatureConfig> DESERT_MINESHAFT = new RSMineshaftStructure(DefaultFeatureConfig.CODEC, RSMineshaftPieces.Type.DESERT, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.spawnrate.desertMineshaftSpawnrate, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.maxHeight.desertMineshaftMaxHeight, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.minHeight.desertMineshaftMinHeight);
+    public static StructureFeature<DefaultFeatureConfig> END_MINESHAFT = new RSMineshaftEndStructure(DefaultFeatureConfig.CODEC, RSMineshaftPieces.Type.END, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.spawnrate.endMineshaftSpawnrate, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.maxHeight.endMineshaftMaxHeight, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.minHeight.endMineshaftMinHeight);
+    public static StructureFeature<DefaultFeatureConfig> NETHER_MINESHAFT = new RSMineshaftStructure(DefaultFeatureConfig.CODEC, RSMineshaftPieces.Type.NETHER, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.spawnrate.netherMineshaftSpawnrate, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.maxHeight.netherMineshaftMaxHeight, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.minHeight.netherMineshaftMinHeight);
+    public static StructureFeature<DefaultFeatureConfig> ICY_MINESHAFT = new RSMineshaftStructure(DefaultFeatureConfig.CODEC, RSMineshaftPieces.Type.ICY, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.spawnrate.icyMineshaftSpawnrate, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.maxHeight.icyMineshaftMaxHeight, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.minHeight.icyMineshaftMinHeight);
+    public static StructureFeature<DefaultFeatureConfig> JUNGLE_MINESHAFT = new RSMineshaftStructure(DefaultFeatureConfig.CODEC, RSMineshaftPieces.Type.JUNGLE, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.spawnrate.jungleMineshaftSpawnrate, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.maxHeight.jungleMineshaftMaxHeight, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.minHeight.jungleMineshaftMinHeight);
+    public static StructureFeature<DefaultFeatureConfig> OCEAN_MINESHAFT = new RSMineshaftStructure(DefaultFeatureConfig.CODEC, RSMineshaftPieces.Type.OCEAN, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.spawnrate.oceanMineshaftSpawnrate, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.maxHeight.oceanMineshaftMaxHeight, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.minHeight.oceanMineshaftMinHeight);
+    public static StructureFeature<DefaultFeatureConfig> SAVANNA_MINESHAFT = new RSMineshaftStructure(DefaultFeatureConfig.CODEC, RSMineshaftPieces.Type.SAVANNA, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.spawnrate.savannaMineshaftSpawnrate, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.maxHeight.savannaMineshaftMaxHeight, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.minHeight.savannaMineshaftMinHeight);
+    public static StructureFeature<DefaultFeatureConfig> STONE_MINESHAFT = new RSMineshaftStructure(DefaultFeatureConfig.CODEC, RSMineshaftPieces.Type.STONE, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.spawnrate.stoneMineshaftSpawnrate, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.maxHeight.stoneMineshaftMaxHeight, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.minHeight.stoneMineshaftMinHeight);
+    public static StructureFeature<DefaultFeatureConfig> SWAMP_OR_DARK_FOREST_MINESHAFT = new RSMineshaftStructure(DefaultFeatureConfig.CODEC, RSMineshaftPieces.Type.SWAMPORDARKFOREST, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.spawnrate.swampAndDarkForestMineshaftSpawnrate, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.maxHeight.swampAndDarkForestMineshaftMaxHeight, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.minHeight.swampAndDarkForestMineshaftMinHeight);
+    public static StructureFeature<DefaultFeatureConfig> TAIGA_MINESHAFT = new RSMineshaftStructure(DefaultFeatureConfig.CODEC, RSMineshaftPieces.Type.TAIGA, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.spawnrate.taigaMineshaftSpawnrate, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.maxHeight.taigaMineshaftMaxHeight, RepurposedStructures.RSAllConfig.RSMineshaftsConfig.minHeight.taigaMineshaftMinHeight);
 
     public static StructureFeature<DefaultFeatureConfig> STONEBRICK_STRONGHOLD = new RSStonebrickStrongholdStructure(DefaultFeatureConfig.CODEC);
     public static StructureFeature<DefaultFeatureConfig> NETHER_STRONGHOLD = new RSNetherStrongholdStructure(DefaultFeatureConfig.CODEC);
@@ -79,15 +79,15 @@ public class RSFeatures {
 
     public static StructureFeature<DefaultFeatureConfig> END_SHIPWRECK = new EndShipwreckStructure(DefaultFeatureConfig.CODEC);
 
-    public static StructureFeature<DefaultFeatureConfig> BADLANDS_VILLAGE = new VillageBadlandsStructure(DefaultFeatureConfig.CODEC);
-    public static StructureFeature<DefaultFeatureConfig> BIRCH_VILLAGE = new VillageBirchStructure(DefaultFeatureConfig.CODEC);
-    public static StructureFeature<DefaultFeatureConfig> DARK_FOREST_VILLAGE = new VillageDarkForestStructure(DefaultFeatureConfig.CODEC);
-    public static StructureFeature<DefaultFeatureConfig> JUNGLE_VILLAGE = new VillageJungleStructure(DefaultFeatureConfig.CODEC);
-    public static StructureFeature<DefaultFeatureConfig> SWAMP_VILLAGE = new VillageSwampStructure(DefaultFeatureConfig.CODEC);
-    public static StructureFeature<DefaultFeatureConfig> MOUNTAINS_VILLAGE = new VillageMountainsStructure(DefaultFeatureConfig.CODEC);
-    public static StructureFeature<DefaultFeatureConfig> GIANT_TAIGA_VILLAGE = new VillageGiantTaigaStructure(DefaultFeatureConfig.CODEC);
-    public static StructureFeature<DefaultFeatureConfig> CRIMSON_VILLAGE = new VillageCrimsonStructure(DefaultFeatureConfig.CODEC);
-    public static StructureFeature<DefaultFeatureConfig> WARPED_VILLAGE = new VillageWarpedStructure(DefaultFeatureConfig.CODEC);
+    public static StructureFeature<DefaultFeatureConfig> BADLANDS_VILLAGE = new VillageBaseStructure(DefaultFeatureConfig.CODEC, new Identifier(RepurposedStructures.MODID + ":village/badlands/town_centers"), 10);
+    public static StructureFeature<DefaultFeatureConfig> BIRCH_VILLAGE = new VillageBaseStructure(DefaultFeatureConfig.CODEC, new Identifier(RepurposedStructures.MODID + ":village/birch/town_centers"), 6);
+    public static StructureFeature<DefaultFeatureConfig> DARK_FOREST_VILLAGE = new VillageBaseStructure(DefaultFeatureConfig.CODEC, new Identifier(RepurposedStructures.MODID + ":village/dark_forest/town_centers"), 6);
+    public static StructureFeature<DefaultFeatureConfig> JUNGLE_VILLAGE = new VillageBaseStructure(DefaultFeatureConfig.CODEC, new Identifier(RepurposedStructures.MODID + ":village/jungle/town_centers"), 8);
+    public static StructureFeature<DefaultFeatureConfig> SWAMP_VILLAGE = new VillageSwampStructure(DefaultFeatureConfig.CODEC, new Identifier(RepurposedStructures.MODID + ":village/swamp/town_centers"), 6);
+    public static StructureFeature<DefaultFeatureConfig> MOUNTAINS_VILLAGE = new VillageBaseStructure(DefaultFeatureConfig.CODEC, new Identifier(RepurposedStructures.MODID + ":village/mountains/town_centers"), 6);
+    public static StructureFeature<DefaultFeatureConfig> GIANT_TAIGA_VILLAGE = new VillageBaseStructure(DefaultFeatureConfig.CODEC, new Identifier(RepurposedStructures.MODID + ":village/giant_taiga/town_centers"), 6);
+    public static StructureFeature<DefaultFeatureConfig> CRIMSON_VILLAGE = new VillageNetherStructure(DefaultFeatureConfig.CODEC, new Identifier(RepurposedStructures.MODID + ":village/crimson/town_centers"), 6);
+    public static StructureFeature<DefaultFeatureConfig> WARPED_VILLAGE = new VillageNetherStructure(DefaultFeatureConfig.CODEC, new Identifier(RepurposedStructures.MODID + ":village/warped/town_centers"), 6);
     public static List<StructureFeature<DefaultFeatureConfig>> OVERWORLD_VILLAGE_LIST = Arrays.asList(
             BADLANDS_VILLAGE,
             BIRCH_VILLAGE,