+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/structures/RSMineshaftEndStructure.java
@@ -17,13 +17,9 @@ import net.minecraft.world.gen.feature.StructureFeature;
 import java.util.List;
 
 
-public class RSMineshaftEndStructure extends AbstractMineshaftStructure {
-    public RSMineshaftEndStructure(Codec<DefaultFeatureConfig> config) {
-        super(config);
-    }
-
-    public double getProbability() {
-        return RepurposedStructures.RSAllConfig.RSMineshaftsConfig.spawnrate.endMineshaftSpawnrate;
+public class RSMineshaftEndStructure extends RSMineshaftStructure {
+    public RSMineshaftEndStructure(Codec<DefaultFeatureConfig> config, RSMineshaftPieces.Type mineshaftType, double probability, int maxHeight, int minHeight) {
+        super(config, mineshaftType, probability, maxHeight, minHeight);
     }
 
     private static final List<Biome.SpawnEntry> MONSTER_SPAWNS = Lists.newArrayList(
@@ -38,34 +34,8 @@ public class RSMineshaftEndStructure extends AbstractMineshaftStructure {
     @Override
     protected boolean shouldStartAt(ChunkGenerator chunkGenerator, BiomeSource biomeSource, long seed, ChunkRandom chunkRandom, int x, int z, Biome biome, ChunkPos chunkPos, DefaultFeatureConfig featureConfig) {
         chunkRandom.setCarverSeed(seed, x, z);
-        double d = (getProbability() / 10000D);
+        double d = (probability / 10000D);
         int landHeight = chunkGenerator.getHeightInGround(x << 4 + 7, z << 4 + 7, Heightmap.Type.WORLD_SURFACE_WG);
         return chunkRandom.nextDouble() < d && (RepurposedStructures.RSAllConfig.RSMineshaftsConfig.misc.barrensIslandsEndMineshafts || landHeight >= 20);
     }
-
-
-    @Override
-    public StructureFeature.StructureStartFactory<DefaultFeatureConfig> getStructureStartFactory() {
-        return RSMineshaftEndStructure.Start::new;
-    }
-
-    public static class Start extends AbstractStart {
-        public Start(StructureFeature<DefaultFeatureConfig> structureIn, int chunkX, int chunkZ, BlockBox mutableBoundingBox, int referenceIn, long seedIn) {
-            super(structureIn, chunkX, chunkZ, mutableBoundingBox, referenceIn, seedIn);
-        }
-
-        @Override
-        public RSMineshaftPieces.Type getMineshaftType() {
-            return RSMineshaftPieces.Type.END;
-        }
-        @Override
-        public int getMaxHeight() {
-            return RepurposedStructures.RSAllConfig.RSMineshaftsConfig.maxHeight.endMineshaftMaxHeight;
-        }
-        @Override
-        public int getMinHeight() {
-            return RepurposedStructures.RSAllConfig.RSMineshaftsConfig.minHeight.endMineshaftMinHeight;
-        }
-    }
-
 }